// learn-redux/src/types/todoTypes.ts
export interface ITodoItem {
  id: string;
  title: string;
  completed: boolean;
}

export enum TodoActionTypes {
  SET_TODO_ITEM = "SET_TODO_ITEM",
  ADD_TODO_ITEM = "ADD_TODO_ITEM",
  SET_ACTION = "SET_ACTION",
  UPDATE_TODO_ITEM = "UPDATE_TODO_ITEM",
  DELETE_TODO_ITEM = "DELETE_TODO_ITEM",
  TOGGLE_TODO_ITEM = "TOGGLE_TODO_ITEM",
  TOGGLE_ALL_TODO_ITEMS = "TOGGLE_ALL_TODO_ITEMS",
  GET_TODO_LIST_FROM_SERVER = "GET_TODO_LIST_FROM_SERVER",
  SET_LOADING = "SET_LOADING",

  FETCH_TODO_START = "FETCH_TODO_START",
  FETCH_TODO_SUCCESS = "FETCH_TODO_SUCCESS",
  FETCH_TODO_FAILURE = "FETCH_TODO_FAILURE",

  ADD_TODO_START = "ADD_TODO_START",
  ADD_TODO_SUCCESS = "ADD_TODO_SUCCESS",
  ADD_TODO_FAILURE = "ADD_TODO_FAILURE",

  UPDATE_TODO_START = "UPDATE_TODO_START",
  UPDATE_TODO_SUCCESS = "UPDATE_TODO_SUCCESS",
  UPDATE_TODO_FAILURE = "UPDATE_TODO_FAILURE",

  DELETE_TODO_START = "DELETE_TODO_START",
  DELETE_TODO_SUCCESS = "DELETE_TODO_SUCCESS",
  DELETE_TODO_FAILURE = "DELETE_TODO_FAILURE",

  TOGGLE_TODO_START = "TOGGLE_TODO_START",
  TOGGLE_TODO_SUCCESS = "TOGGLE_TODO_SUCCESS",
  TOGGLE_TODO_FAILURE = "TOGGLE_TODO_FAILURE",

  TOGGLE_ALL_START = "TOGGLE_ALL_START",
  TOGGLE_ALL_SUCCESS = "TOGGLE_ALL_SUCCESS",
  TOGGLE_ALL_FAILURE = "TOGGLE_ALL_FAILURE",
}